+++ src/krit/Utils.cpp
@@ -5,26 +5,26 @@ namespace krit {
 std::random_device rd;
 std::mt19937 rng(rd());
 
-template <> double smoothStep<0>(double n) {
+template <> float smoothStep<0>(float n) {
     return pow(n, 2) * (3 - 2 * n);
 }
 
-double maybeLerp(double t, double f(double)) {
+float maybeLerp(float t, float f(float)) {
     if (f) {
         t = f(t);
     }
     return t;
 }
 
-double noChange(double t) {
+float noChange(float t) {
     return 0;
 }
 
-double easeInOut(double t) {
+float easeInOut(float t) {
     return pow(2 * t - 1, 2);
 }
 
-double easeOutBounce(double t) {
+float easeOutBounce(float t) {
     if (t < (1/2.75)) {
         return 7.5625*t*t;
     } else if (t < (2/2.75)) {
